All API

HttpClient.remove_error_buffer — Method

Remove link to error buffer. Therefore buffer may be safely deleted now.

HttpClient.set_body — Method

Use for post or put request. Body is String, Vector{UInt8} or Nothing.

HttpClient.set_connect_only — Method

Must be set for web socket.

HttpClient.set_delete — Method

Use for delete request.

HttpClient.set_error_buffer — Method

Create vector of bytes and set it as error buffer.

HttpClient.set_follow_location — Method

Support redirection.

HttpClient.set_headers — Method

Return pointer, which must be freed.

HttpClient.set_interface — Method

Use nothing over empty string.

HttpClient.set_put — Method

Use for put requests.

HttpClient.set_response — Method

Initialize callback and buffer for response message.

HttpClient.set_url — Method

set_url(easy_handle, url, query) -> curl_url, full_url

Make full url from url and query, then set it to easy handle. curl_url must be freed after perform.

HttpClient.get_headers — Method

Use after perform

HttpClient.get_http_code — Method

Use after perform

HttpClient.join_messages — Method

join_messages(code, error_buffer)

Join message associated with code and specific message from error_buffer.

HttpClient.@curlok — Macro

curlok(exp)

Raise error, if result of exp is not CURLE_OK.

HttpClient.curl_write_cb — Method

callback function to receive data