Web sockets

HttpClient.Connection

Base.close

Base.isopen

HttpClient.open_connection

HttpClient.receive

HttpClient.receive_any

HttpClient.receive_pong

HttpClient.send

HttpClient.send_ping

HttpClient.send_pong

HttpClient.websocket

HttpClient.Connection — Type

Connection

A connection to server via web socket protocol.

See also send, receive, isopen

Usage

connection = open_connection(url; ...)
# your code
close(connection)

websocket(url; ...) do connection
    # your code
end

Base.close — Function

close(c::Connection)

Base.isopen — Method

isopen(c::Connection) -> Bool

HttpClient.open_connection — Method

open_connection(url; <keyword arguments>) -> Connection

Connect to a web socket server.

Keyword Arguments

headers = Pair{String, String}[]. Any iterable container of Pair{String, String}.

query = nothing. A string or pairs of strings. Automatic url encoding.

interface = nothing. Outgoing network interface. Interface name, an IP address, or a host name.

connect_timeout = 60. Abort initial request after timeout is reached.

proxy = nothing. Unsupported.

verbose = false. Show libcurl verbose messages.

isdeflate = false. Decompress messages from server using deflate protocol.

HttpClient.receive — Method

receive(connection) -> message

Receive message from web socket while handling all control messages. Close connection on error.

To monitor control messages, you can display debug messages via ENV["JULIA_DEBUG"] = "HttpClient"

HttpClient.receive_any — Method

receive_any(connection) -> message, message_type

Receive any message from server and don't perform any control actions. Possible message types are text, binary, ping, pong, close, missing.

User is responsible for handling control messages.

HttpClient.receive_pong — Function

receive_pong(connection, message = "foo") -> ispong, data

Try to receive pong message. Warning if receive not pong. Error if receive wrong message.

HttpClient.send — Function

send(connection, message; flags)

Send message to web socket. Close connection on error.

Use String as message for text and Vector{UInt8} for binary data. flags are set automatically based on message type. See curl docs for possible values.

Example

HttpClient.send(connection, "Hello!")

HttpClient.send_ping — Function

send_ping(connection, message = "foo")

Send ping message to a server.

WARNING empty message may close connection.

HttpClient.send_pong — Function

send_pong(connection, message = "foo")

Send pong message to a server.

WARNING empty message may close connection.

HttpClient.websocket — Method

websocket([handle,] url; <keyword arguments>) -> Connection

Connect to a web socket server, run handle on connection, then close connection.

Keyword Arguments

headers = Pair{String, String}[]. Any iterable container of Pair{String, String}.

query = nothing. A string or pairs of strings. Automatic url encoding.

interface = nothing. Outgoing network interface. Interface name, an IP address, or a host name.

connect_timeout = 60. Abort initial request after timeout is reached.

proxy = nothing. Unsupported.

verbose = false. Show libcurl verbose messages.

isdeflate = false. Decompress messages from server using deflate protocol.

Example

websocket(url; ...) do connection
    # your code
end

f(connection) = # your code
websocket(f, url; ...)